Add support for specifying particular card versions in a cube
A cube card can now be written as "Abrade (CMR)" or "Abrade (CMR 659)"
to specify the exact image that should be displayed. To enable this,
the cubable card database has been expanded to include a map from card
names to set codes to version numbers, as well as a "default" version
for each card that's based on the same logic that's historically been
used to select images.