Fix adding brackets to blank and generic-only mana costs
*Explicitly convert manacosts to strings before passing them to addManaCostBrackets
*Test manacost strings against manaSymbolRegExp before attempting to match mana symbols within them. If the regExp tests false, skip adding brackets and just pass the string back.

This commit corrects two errors: the first occurred when attempting to add brackets to manacost values that had only generic mana symbols (such as '1'). Because the parser was interpreting these as numbers and not as strings, the String.prototype.match() method was failing and aborting xml import.

The second was in the case of cards with empty manacost values (such as lands.) In these cases the regex match created a null manaSymbolArray, which could not be mapped and also caused import to fail.