Mega pr (#1486)
* Fix manaCost and convertedManaCost parsing in uploaded set xml 

*Label cmc correctly as convertedManaCost in set json files generated by parser.js from uploaded Cockatrice set xml files
*Use a regex to match mana symbols in manaCost and add brackets to them in set json files generated by parser.js from uploaded Cockatrice set xml files

* Update parser.js spec accordingly

* Preserve split card indicator in manaCost parsing

This commit moves the manaSymbolRegExp definition to outside the AddManaCostBrackets function and alters it to also catch the split card indicator (//). If the split card indicator is found, it's written to the final output wrapped in spaces instead of curly brackets.

* Update parser.js spec to reflect preservation of split card indicator

* Fix adding brackets to blank and generic-only mana costs

*Explicitly convert manacosts to strings before passing them to addManaCostBrackets
*Test manacost strings against manaSymbolRegExp before attempting to match mana symbols within them. If the regExp tests false, skip adding brackets and just pass the string back.

This commit corrects two errors: the first occurred when attempting to add brackets to manacost values that had only generic mana symbols (such as '1'). Because the parser was interpreting these as numbers and not as strings, the String.prototype.match() method was failing and aborting xml import.

The second was in the case of cards with empty manacost values (such as lands.) In these cases the regex match created a null manaSymbolArray, which could not be mapped and also caused import to fail.

* add additional favicons + viewport meta

* allow semver minor upgrades, bump fuse.js major

* Improve & clarify style as suggested by @mixmix

* fix flipped cards in card preview. add flip transition

* round card corners

Co-authored-by: Iniquit <iniquitwrites@gmail.com>
Co-authored-by: Jono Kemball <Noojuno@users.noreply.github.com>
Co-authored-by: Zach H <zahalpern+github@gmail.com>